Part4/pyproject.toml
[project]
name = "part2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.13.0",
    "pytest>=8.3.5",
]
